requirements.txt
absl-py
beautifulsoup4
selenium
webdriver-manager